Connected backend to db and wrote first route
Installed the sqlite & sqlite3 libraries to connect to the db.
We need both because it allows us to use async/await which is much
easier than callbacks. Also I made a .gitignore file because we
don't want to commit the node_modules folder on the backend.
Lastly, I wrote a route similar to the one on the prototype. It
gives the user the classes available in a given semester and the
total amount of space in all the classes. The specific semester
used is hardcoded in but that will be changed in the future.