Created route that inputs new semester
I created a route that takes in a semester as a json object and
puts it into the database. That way, if we wanted to put requests
for a future semester, we could do so. If the semester already
exists, it'll send a failure message. Otherwise, it'll create it
and send a success message. It does not check for case so it's
possible for fall2022 and Fall2022 to be 2 different semesters.