Updated DB Schema
I changed the schema so that the finalized column is in the Semester
table instead of the Assignments table. I did this bc, you would
only finalize TA assignments once every TA got assigned or the
classes ran out of space for a specific semester. In other words,
the finalized column relates more to a specific semester than it
does to individual assignments.