Implemented assigned_student object for reranking route
Before this, it was a static object. There's a bug though bc when
assigned_student gets sent to the backend it's empty. It's weird
though bc I console.log the object itself on the front end and it
looks normal. But when I console.log assigned_student.id on the
front-end, it's undefined.